requirements.txt
fastapi
uvicorn
jinja2
pandas
fpdf2
python-multipart
pydantic